+++ src/components/themes/BaseTheme.h
@@ -169,9 +169,10 @@ class BaseTheme {
                          PopupShip ship = PopupShip::Blocking) const;
   virtual void fillPopupProgress(const GfxRenderer& renderer, const Rect& layout, const int progress) const;
   virtual void drawStatusBar(GfxRenderer& renderer, const float bookProgress, const int currentPage,
-                             const int pageCount, std::string title, const int paddingBottom = 0,
-                             const bool isStarred = false, const std::string& printedPageLabel = std::string(),
-                             const bool fillMargin = true, const bool pageCountApproximate = false) const;
+                             const int pageCount, const std::string& bookTitle, const std::string& chapterTitle,
+                             const int paddingBottom = 0, const bool isStarred = false,
+                             const std::string& printedPageLabel = std::string(), const bool fillMargin = true,
+                             const bool pageCountApproximate = false) const;
   virtual void drawHelpText(const GfxRenderer& renderer, Rect rect, const char* label) const;
   virtual void drawTextField(const GfxRenderer& renderer, Rect rect, const int textWidth, bool cursorMode = false,
                              int contentStartX = 0, int contentWidth = 0) const;